# Backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "REVENUE HUB",
        "description": "Below is the API endpoint for the Revenue Hub",
        "contact": {
            "email": "darius@matulionis.lt"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0",
        "x-logo": {
            "url": "https://via.placeholder.com/190x90.png?text=L5-Swagger"
        }
    },
    "paths": {
...
        "/api/staff": {
            "get": {
                "tags": [
                    "Staff"
                ],
                "summary": "Get all staffs",
                "description": "Show list of all staffs",
                "operationId": "getStaffs",
                "responses": {
                    "200": {
                        "description": "List of all staffs",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/StoreUserResource"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "No staff found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Staff"
                ],
                "summary": "Register New Staff",
                "description": "This allow the MD to create new staff member",
                "operationId": "registerStaff",
                "requestBody": {
                    "description": "Input staff details to register",
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "phone",
                                    "zone",
                                    "role_id"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "Staff Fullname",
                                        "type": "text",
                                        "example": "Revenue Hub"
                                    },
                                    "email": {
                                        "description": "Staff Email",
                                        "type": "text",
                                        "example": ""
                                    },
                                    "phone": {
                                        "description": "Staff Phone",
                                        "type": "text",
                                        "example": ""
                                    },
                                    "zone": {
                                        "description": "Staff Zone",
                                        "type": "text",
                                        "example": ""
                                    },
                                    "role_id": {
                                        "description": "Staff Roles",
                                        "type": "string",
                                        "enum": [
                                            "2",
                                            "3",
                                            "4"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Registeration Successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/StoreUserResource"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "All fields are required",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "All fields are required"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "402": {
                        "description": "An error occurred",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "An error occurred"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "StoreUserResource": {
                "title": "Store Staff Resource",
                "description": "Store Staff Resource",
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "api_key": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer token)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}